refactor(bus): migrate the whole probe fleet to the float64 value bus (#271/#407)
DataPoint.Value is float64 (#407 bus migration, re-applied here). Every probe
built on the old float32 bus is converted: DataPoint values widen from float32
to float64 across all ~43 probe packages and their tests. This removes the
24-bit-mantissa precision loss that stair-stepped large counters (SNMP/IPMI/
proxmox/uint32 modbus registers) above 2^24 — the bug #271 targets.

modbus keeps float32 only where it must: the IEEE-754 32-bit wire decode
(math.Float32frombits) and the Scale config factor; readRegister now returns
float64 so a uint32 register value no longer round-trips through float32.
Duplicate float64 type-switch cases introduced by the mechanical widening
(params.go, tomcat) were collapsed.

make test green, go vet clean. Supersedes the float32-bus exclusion in the
0.3.0 integration branch.

+++ internal/agent/probes/activemq/activemq_probe.go
@@ -336,7 +336,7 @@ func (p *activemqProbe) collectBroker(ctx context.Context, now time.Time) ([]dat
 
 	points = append(points, data_store.DataPoint{
 		Name:      "activemq.producer.count",
-		Value:     float32(producerCount),
+		Value:     float64(producerCount),
 		Timestamp: now,
 		Tags:      brokerTags,
 	})
@@ -349,7 +349,7 @@ func (p *activemqProbe) collectBroker(ctx context.Context, now time.Time) ([]dat
 		}
 		points = append(points, data_store.DataPoint{
 			Name:      a.metric,
-			Value:     float32(v),
+			Value:     float64(v),
 			Timestamp: now,
 			Tags:      brokerTags,
 		})
@@ -363,7 +363,7 @@ func (p *activemqProbe) collectBroker(ctx context.Context, now time.Time) ([]dat
 		}
 		points = append(points, data_store.DataPoint{
 			Name:      a.metric,
-			Value:     float32(v * a.scale),
+			Value:     float64(v * a.scale),
 			Timestamp: now,
 			Tags:      brokerTags,
 		})
@@ -505,7 +505,7 @@ func (p *activemqProbe) collectOneDestination(
 		}
 		points = append(points, data_store.DataPoint{
 			Name:      a.metric,
-			Value:     float32(v),
+			Value:     float64(v),
 			Timestamp: now,
 			Tags:      destTags,
 		})
